requirements.txt
requests
pandas
pyyaml


